AppOpsService: Don't log stack traces for isolated processes
AppOpsService logs a stack trace if the package and the UID
of the call don't match, which includes a stack trace. Since
an isolated process is running under UID that is different
from that of the app, it is expected that the package and
the UID don't match - this is by design. In this case, a
stack trace is not helpful and spams the logs. Therefore,
for isolated processes we log a short message.

Test: isloated process logs have no stack trace

Upstream from Meta

Change-Id: Ife7bff103c3a4d1823de92461248451ff313de0e
Signed-off-by: Anushek Prasal <anushekprasal@gmail.com>